% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calc_JK}
\alias{calc_JK}
\title{calc_JK}
\usage{
calc_JK(EST, LOO_EST, alpha = 0.05)
}
\arguments{
\item{EST}{A numeric vector of parameter estimates}

\item{LOO_EST}{A numeric matrix of parameter estimates
where columns correspond to each parameter and
rows correspond to each leave one out estimate}

\item{alpha}{A numeric value for constructing
(1 - alpha) * 100\% confidence intervals}
}
\value{
A list of numeric jackknife summary mean 
and confidence intervals
}
\description{
calc_JK
}
