% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{chk_threads}
\alias{chk_threads}
\title{chk_threads}
\usage{
chk_threads(NN, ncores)
}
\arguments{
\item{NN}{A positive integer for total iterations
to loop over}

\item{ncores}{A positive integer for number of threads}
}
\value{
An integer for number of threads.
}
\description{
chk_threads
}
