% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packs.R
\name{chkInst_PACK}
\alias{chkInst_PACK}
\title{chkInst_PACK}
\usage{
chkInst_PACK(PACK)
}
\arguments{
\item{PACK}{A character string for a package name}
}
\value{
Boolean for \code{TRUE} if package installed
and \code{FALSE} if package is not installed or located
}
\description{
Check package is installed
}
