% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{collapse_var}
\alias{collapse_var}
\title{collapse_var}
\usage{
collapse_var(ORIG_VAR, ORIG_VALUES, NEW_VALUE)
}
\arguments{
\item{ORIG_VAR}{The input vector}

\item{ORIG_VALUES}{A subset of values from
the input vector to be collapsed}

\item{NEW_VALUE}{The new value to replace
\code{ORIG_VALUES} in \code{ORIG_VAR}}
}
\description{
collapse_var
}
