% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_species_attribute.R
\name{create_species_attribute}
\alias{create_species_attribute}
\title{Create species attribute for SMART Desktop 7's Data Models}
\usage{
create_species_attribute(
  language = "en",
  species_attribute_key = "species_smarter",
  species_in_language = "species",
  first_tax_level = first_tax_level,
  species_key = species_key,
  species_label = species_label
)
}
\arguments{
\item{language}{The abbreviation of the language of the attribute as a string.
It must be compatible with the it has to be consistent with the languages
available in the data model in SMART Desktop. If your Conservation Area has
as language "english" and "laotian", then the argument 'language' can be
either "en" or "lo" but not "fr" (french). To find the abbreviations check
the options in your Conservation Area Data Model on the top-left corner and
click on the downward arrow.}

\item{species_attribute_key}{A string that will become the key of the species
attribute. It should not have weird characters. The function takes care of
uppercase letters and white spaces.}

\item{species_in_language}{A string that will become the label of the tree
attribute}

\item{first_tax_level}{A character vector to provide the keys and labels of
the fist taxonomic level of the species attribute.}

\item{species_key}{A character vector to provide the keys of the species within
each taxonomic group of the first level of the tree. Vector elements should not
have weird characters. The function takes care of uppercase letters and white
spaces.}

\item{species_label}{A character vector to provide the labels of the species
within each taxonomic group of the first level of the tree (e.g., common local
names).}
}
\value{
xml file named "species_smarter_attribute.xml" saved in your working directory.
This is the "Data Model" with just the species attribute to be merged to the
Conservation Area's Data model and get the species attribute available.
}
\description{
Function to create a data model that only has a species attribute as a tree.
Then this data model with just the species tree can be merged to the Conservation
Areas' data model and it becomes available to be added to the Categories.
}
\details{
Basically the function 'create_species_function' creates a Data Model that only
has a species attribute as a tree with two taxonomic levels. Then this species
attribute-only data model can be merged into the user's Data Model and the
attribute becomes available to be added to the Categories. The species tree
attribute can be flattened to a list in SMART Desktop.
The output of the function is an xml file named "species_smarter_attribute.xml"
saved in your working directory. To add the attribute to your data model, open
the Conservation Area, open the Data Model, find the "Merge Data Model" button
in the bottom left and select the "species_smarter_attribute.xml" file. Add the
attribute to the Category or Categories (search for the attribute usig the key
provided in 'species_attribute_key').
}
\examples{
first_tax_level<-c(rep('Aves', 3), rep("Mammalia", 3))

species_key=c('Circus melanoleucos', 
              'Spilornis cheela', 
              'Dendrocygna javanica', 
              'Bos gaurus', 
              'Bos javanicus', 
              'Capricornis milneedwardsii maritimus')
              
species_label=c('Pied harrier',
  'Crested serpent eagle',
   'Lesser whistling duck',
   'Wild water buffalo',
   'Red muntjac',
   'Large-antlered Muntjac')
   
create_species_attribute(
language="en", 
species_attribute_key="species_smarter",
species_in_language="species",
first_tax_level=first_tax_level,
species_key=species_key,
species_label=species_label)
}
