% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_from_connect.R
\name{data_from_connect}
\alias{data_from_connect}
\title{Run a SMART query in SMART Connect and open the output in an R session.}
\usage{
data_from_connect(
  server_url,
  user,
  password,
  name_conservation_area,
  query_name,
  type_output,
  directory,
  date_filter = "waypointdate",
  start_date = NULL,
  end_date = NULL,
  srid = 4326,
  UTM_zone = NULL
)
}
\arguments{
\item{server_url}{A string with the URL of the SMART Connect server
(e.g., "https://wcshealth.smartconservationtools.org/server" )}

\item{user}{A string with corresponding SMART Connect username
(e.g., "connect_username")}

\item{password}{A string with the corresponding SMART Connect user's password
(e.g., "my_connect_password")}

\item{name_conservation_area}{A string he name of the Conservation Area
holding the query}

\item{query_name}{A string with the name of the SMART query to run
(e.g., "query_name")}

\item{type_output}{Either "csv" or "shp" to load the query data as
a tibble or sf object. Not all queries have spatial data. See details.}

\item{directory}{a path to a directory with the proper permissions to write
a temp file with the spatial data}

\item{date_filter}{A string with the field establishing the start and
end dates of the data of the query's output. Default is "waypointdate".
See details.}

\item{start_date}{A string with the start date filter with the format
YYYY-MM-DD (e.g, "2000-01-01"). Default is NULL.}

\item{end_date}{A string with the end date filter with the format
YYYY-MM-DD (e.g, "2000-01-01"). Default is NULL.}

\item{srid}{An integer with the EPSG code of the projection for a spatial
output (when "shp" is selected). Default is 4326.}

\item{UTM_zone}{An integer to provide the Zone if a Universal Transverse
Mercator is the srid input. From 1 to 60.}
}
\value{
A tibble or sf object with the data returned by the query run
in SMART Connect. Columns and rows are returned as defined in the source
query. If summary queries are requested, the output does not have adequate
column names and only the very fist header is considered a column. In any case
if you are loading the data in R it is because you want to summarize data here.
This function should be use for queries returning raw data (patrol queries,
patrol observation queries, etc.).
}
\description{
A function to load data from queries available in SMART Connect in specific
Conservation Areas. Depending on the type of query, it is possible to load
the data as a spatial object. To learn more about the properties of your
queries, use the function "query_info".
}
\details{
Not all SMART queries have spatial information included. Different query
types have different properties including day filter options. To learn about
the Conservation Areas available in SMART Connect, the queries available per
Conservation Area, and their properties such as query type, spatial
information availability (the type_output parameter of this function), and
date filter options (the date_filter parameter of this function) use the
"query_info" function first. Avoid empty query folders.
}
\examples{

server_url <- Sys.getenv("EXAMPLE_CONNECT_URL")
user <- Sys.getenv("EXAMPLE_CONNECT_USERNAME")
password <-Sys.getenv("EXAMPLE_CONNECT_PASSWORD")
directory <- Sys.getenv("EXAMPLE_DIRECTORY")
                           
data<-data_from_connect(server_url = server_url,
                           user = user, 
                           password = password,
                           name_conservation_area = "Example Conservation Area [SMART]",
                           query_name = "Patrol_query",
                           type_output = "shp",
                           directory = directory,
                           date_filter="waypointlastmodified",
                           start_date="2020-01-01", #YYYY-MM-DD
                           end_date="2023-07-31", #YYYY-MM-DD
                         srid=4326,
                         UTM_zone=NULL)

}
\seealso{
\code{\link[=query_info]{query_info()}} to learn about the Conservation
Areas available in SMART Connect, the queries in each Conservation Area, the
type of each query, spatial data in the queries, and the options available to
filter by date for each query.
,
}
