% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_conf_model.R
\name{flat_conf_model}
\alias{flat_conf_model}
\title{Flat table of a Configurable Model in a specified language}
\usage{
flat_conf_model(path_conf_model, language_interest = "en", only_active = T)
}
\arguments{
\item{path_conf_model}{the path to the xml file ending in "/xml_file_name.xml"}

\item{language_interest}{a string code with the language of interest. "en"
is the default value. If configurable model has more than one language, use
this function once per language. The value for this argument must be consistent
with one of the languages available for the configurable model.}

\item{only_active}{a boolean to choose if the flat table should include only
active attributes, tree roots, and options (TRUE) or everything (FALSE)}
}
\value{
A tibble. The tibble has the following
columns:
"cat_key_level_n", "cat_label_level_n", "att_key", "att_label", "att_type",
"att_active", "root_key", "root_label", "root_active", "option_key", and
"option_label", "option_active". The "cat", "att", "root", and "option"
strings in the column headers refer to Category, Attribute, Tree and Option
data, respectively. The "key", "label", and "option" strings refer to data
regarding the keys and the labels, The "level_n" refers to the
level of the nested Category. Keys do not change across languages.
}
\description{
Function to create a flat table from a SMART Configurable Model downloaded
from a Conservation Area in SMART Desktop exported as xml. The output has all
categories, attributes (list, multilist, trees, and others), and options.
Keys and labels for each item are provided. All inactive categories, attributes,
and options are NOT included in the result.
}
\details{
Currently, it is not possible to obtain the configurable model directly from
SMART Connect, so the Configurable Model must be exported from the Conservation
Areas in SMART Desktop as an xml file to a known destination in your computer.
Then provide the path to the xml file in your local computer in the 'path_conf_model'
attribute of this function.
To export a Configurable Model, open SMART Desktop and the corresponding
Conservation Area. Go to the Conservation Area tab and click on "Configurable
Model". Select the Configurable Model you would like to present as a flat
table from the list and then click on "Export". Choose the location to save the
xml file in your computer. Then run this function providing the corresponding
path.

For now, tree attributes of the configurable model are assumed to have roots
and options (two levels top). Otherwise you will get a wrong output for tree
attributes.
}
\examples{
path_conf_model <- system.file("extdata", "example_configurable_model.xml", package = "SMARTeR")

flat_conf_model(
path_conf_model=path_conf_model, 
language_interest="en",
only_active=T)

}
