% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{make_dummy}
\alias{make_dummy}
\title{make_dummy}
\usage{
make_dummy(x)
}
\arguments{
\item{x}{A numeric or character vector
to convert to a dummy matrix}
}
\value{
A binary indicator matrix of ones and zeros
}
\description{
Construct a dummy-coded matrix 
for a single variable
}
