% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_menu.R
\name{make_menu}
\alias{make_menu}
\title{make_menu}
\usage{
make_menu(PROMPT, OPTS, INDENT = "   ")
}
\arguments{
\item{PROMPT}{A character string prompt to the user}

\item{OPTS}{A character vector where elements contain
a number, then a closing parentheses, then the option 
value}

\item{INDENT}{A character string for the amount 
of indentation from the left margin}
}
\value{
Character string of user's response
}
\description{
Constructs an interactive menu
 for the user
}
