% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{name_change}
\alias{name_change}
\title{name_change}
\usage{
name_change(DATA, ORIG_NAME, NEW_NAME)
}
\arguments{
\item{DATA}{A matrix or data.frame}

\item{ORIG_NAME}{A single character column name to alter}

\item{NEW_NAME}{A single character to replace \code{ORIG_NAME}}
}
\value{
An updated data.frame
}
\description{
Substitute a column name of a matrix 
or data.frame with a new name
}
