% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{print_latex_table}
\alias{print_latex_table}
\title{print_latex_table}
\usage{
print_latex_table(
  DATA,
  repeat_VARS = NULL,
  my_align = NULL,
  add_table = FALSE,
  fontsize = NULL,
  caption = NULL,
  label = NULL,
  midrule1 = NULL,
  latex_comment = NULL,
  ...
)
}
\arguments{
\item{DATA}{A matrix or data.frame to
present as a latex table}

\item{repeat_VARS}{A string vector of colnames
to avoid repeating values within a column}

\item{my_align}{A string containing letters "l",
"r", or "c" for left, right, and center alignment}

\item{add_table}{Boolean set to TRUE to
enclose tabular environment with table environment}

\item{fontsize}{Defaults to NULL to not specify
a fontsize. Otherwise, possible values are
"tiny", "footnotesize", "small", "normalsize",
"large", "Large", "LARGE", "huge","Huge"}

\item{caption}{A string to include a table caption}

\item{label}{A string to represent a latex table label}

\item{midrule1}{Default is set to NULL}

\item{latex_comment}{Add a latex comment above 
the table for notes}

\item{...}{arguments passed to cat}
}
\value{
No return value
}
\description{
print_latex_table
}
