% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{smart_boxplot}
\alias{smart_boxplot}
\title{smart_boxplot}
\usage{
smart_boxplot(MAT, mar_down = 8, srt = 45, ...)
}
\arguments{
\item{MAT}{A numeric matrix of columns to 
plot as boxplots}

\item{mar_down}{A positive numeric value to 
allow space below the x-axis for labels}

\item{srt}{A numeric value to control the 
angle of x-axis labels}

\item{...}{arguments passed to boxplot}
}
\description{
smart_boxplot
}
