% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{smart_colors}
\alias{smart_colors}
\title{smart_colors}
\usage{
smart_colors(nn, alpha = 1, overwrite = FALSE)
}
\arguments{
\item{nn}{A positive integer greater than 
or equal to 2}

\item{alpha}{A positive numeric value less 
than or equal to one}

\item{overwrite}{Boolean If nn = 2, setting to 
FALSE will force colors to be white or black}
}
\value{
No return value.
}
\description{
smart_colors
}
