% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{smart_compMATs}
\alias{smart_compMATs}
\title{smart_compMATs}
\usage{
smart_compMATs(
  MAT1,
  MAT2 = NULL,
  which_range = NULL,
  xlab,
  ylab,
  show_corr = TRUE,
  show_plot = FALSE,
  main = NULL,
  vec_col = NULL,
  ...
)
}
\arguments{
\item{MAT1}{A numeric matrix}

\item{MAT2}{A second numeric matrix of columns
comparable to MAT1. Default is set to \code{NULL}
resulting in histograms plotted for columns of MAT1}

\item{which_range}{Default is set to \code{NULL}
to calculate data ranges. Otherwise if set to "01",
will enforce minimum 0 and maximum 1. If set to a
numeric vector of two elements, will enforce the range.}

\item{xlab}{A string for x-axis label}

\item{ylab}{A string for y-axis label}

\item{show_corr}{Boolean set to TRUE to print
Pearson and Spearman correlations}

\item{show_plot}{Boolean set to TRUE to plot
comparison of two matrices}

\item{main}{A string for the plot title}

\item{vec_col}{A vector of colors to color scatter 
plot points}

\item{...}{arguments passed to plot}
}
\description{
smart_compMATs
}
