% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{smart_df}
\alias{smart_df}
\title{smart_df}
\usage{
smart_df(...)
}
\arguments{
\item{...}{arguments passed to data.frame}
}
\value{
A data.frame
}
\description{
Construct data.frame with \code{data.frame()}
but sets \code{stringsAsFactors} to \code{FALSE}.
}
