% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartr.R
\name{smart_digits}
\alias{smart_digits}
\title{smart_digits}
\usage{
smart_digits(x, digits = 2)
}
\arguments{
\item{x}{A numeric vector formatted
to have consistently rounded values}

\item{digits}{A positive integer
to regulate the number of digits to
round to}
}
\description{
smart_digits
}
