% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{smart_dots}
\alias{smart_dots}
\title{smart_dots}
\usage{
smart_dots(wait = 300, num_dots = 30)
}
\arguments{
\item{wait}{A number of seconds to wait before printing "."}

\item{num_dots}{The number of dots to print before
printing a message}
}
\value{
No return value.
}
\description{
Prints a series of dots every
few seconds
}
