% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{smart_heatmap}
\alias{smart_heatmap}
\title{smart_heatmap}
\usage{
smart_heatmap(
  MAT = NULL,
  DIST = FALSE,
  main = "",
  width = NULL,
  height = NULL,
  GRID = NULL,
  clustRC = c(TRUE, TRUE),
  nodePar_col = NULL,
  nodePar_row = NULL,
  mar = 2,
  cex.main = 2,
  rowData = NULL,
  colData = NULL,
  make_key = TRUE,
  vec_cols = NULL
)
}
\arguments{
\item{MAT}{A numeric matrix of values}

\item{DIST}{Boolean set to TRUE to treat MAT as distance
matrix. Otherwise, function can perform row/column clustering}

\item{main}{A string for the overall heatmap title}

\item{width}{NULL}

\item{height}{NULL}

\item{GRID}{NULL}

\item{clustRC}{NULL}

\item{nodePar_col}{NULL}

\item{nodePar_row}{NULL}

\item{mar}{NULL}

\item{cex.main}{NULL}

\item{rowData}{NULL}

\item{colData}{NULL}

\item{make_key}{NULL}

\item{vec_cols}{NULL}
}
\description{
smart_heatmap
}
