% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{smart_hist}
\alias{smart_hist}
\title{smart_hist}
\usage{
smart_hist(x, freq = FALSE, dens = TRUE, main = "", ...)
}
\arguments{
\item{x}{A numeric vector}

\item{freq}{Boolean set to \code{FALSE}
to plot density on y-axis. Otherwise
\code{TRUE} to plot frequencies}

\item{dens}{Boolean set to \code{TRUE}
to overlay kernel density}

\item{main}{String for plot title}

\item{...}{arguments passed to hist}
}
\description{
smart_hist
}
