% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{smart_mkdir}
\alias{smart_mkdir}
\title{smart_mkdir}
\usage{
smart_mkdir(input_dir)
}
\arguments{
\item{input_dir}{A full path name for 
a directory to create}
}
\description{
Create directory if it does not exist
}
