% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{smart_names}
\alias{smart_names}
\title{smart_names}
\usage{
smart_names(MAT, ROW = NULL, COL = NULL)
}
\arguments{
\item{MAT}{A matrix}

\item{ROW}{A vector of length equal to \code{nrow(MAT)}}

\item{COL}{A vector of length equal to \code{ncol(MAT)}}
}
\description{
smart_names
}
