% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packs.R
\name{smart_pack}
\alias{smart_pack}
\title{smart_pack}
\usage{
smart_pack(
  pack,
  repo = "CRAN",
  src_fn = "",
  uname = "",
  vers = "",
  pack_dir = "",
  all_packs = NULL
)
}
\arguments{
\item{pack}{A string for the package name}

\item{repo}{A string that takes values
"CRAN" for CRAN, "BIOC" for Bioconductor,
"source" for a package installed from a 
compressed file, "cCRAN" for contributed packages,
"aCRAN" for archived packages, "github" for 
installing packages directly from github, or
"local" for installing a package from a local directory.}

\item{src_fn}{If repo = "source", specify 
the full path to the source file}

\item{uname}{If \code{repo = "github"}, specify
the github username or organization containing the repository}

\item{vers}{If repo = "cCRAN" or "aCRAN",
specify the specific package version to install}

\item{pack_dir}{If repo = "local", specify
the full path to the local package directory}

\item{all_packs}{Defaults to \code{NULL}. Otherwise
a character vector of installed packages}
}
\description{
Check if package installed or 
install package
}
