% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packs.R
\name{smart_pack_versions}
\alias{smart_pack_versions}
\title{smart_pack_versions}
\usage{
smart_pack_versions(pack, repo = "CRAN")
}
\arguments{
\item{pack}{A string for the package name}

\item{repo}{A string that takes values
"CRAN", "aCRAN", and "cCRAN" for combining
options "aCRAN" and "cCRAN". "cCRAN" refers
to contributed packages. "aCRAN" refers
to archived packages.}
}
\description{
Return all associated package versions
}
