% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packs.R
\name{smart_packDeps}
\alias{smart_packDeps}
\title{smart_packDeps}
\usage{
smart_packDeps(
  cran_packs = NULL,
  bioc_packs = NULL,
  github_packs = NULL,
  local_packs = NULL,
  pandoc = NULL,
  build_vign = FALSE
)
}
\arguments{
\item{cran_packs}{A character string of CRAN package names.}

\item{bioc_packs}{A character string of Bioconductor package names.}

\item{github_packs}{A character string of GitHub repositories of the form
organization/repo or username/repo.}

\item{local_packs}{A character string of local full path names R packages}

\item{pandoc}{Defaults to \code{NULL}. If specified, it is a character string
for the full path of pandoc.}

\item{build_vign}{Boolean for building R package vignettes if available.}
}
\description{
Install package with accompandied dependencies and load
}
