% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packs.R
\name{smart_prepPack}
\alias{smart_prepPack}
\title{smart_prepPack}
\usage{
smart_prepPack(
  pack_dir = NULL,
  pandoc = NULL,
  make_vign = FALSE,
  cran = TRUE,
  build_dir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{pack_dir}{Defaults to \code{NULL}, assumes \code{getwd()} is the package directory.
Otherwise, a character string for the local package directory.}

\item{pandoc}{Defaults to \code{NULL}. If specified, it is a character string
for the full path of pandoc.}

\item{make_vign}{Boolean set to \code{FALSE} by default. If \code{TRUE},
attempts to compile available vignettes when checking the package.}

\item{cran}{Boolean set to \code{TRUE} by default. If \code{FALSE},
don't check CRAN specifications.}

\item{build_dir}{Defaults to \code{NULL}. Otherwise specify a
character string for the path to build the package tar.gz file under.}

\item{verbose}{Boolean set to \code{TRUE} by default. Provides
verbose output of package preparation.}
}
\description{
Test, check, install, and maybe build a package
}
