% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartr.R
\name{smart_progress}
\alias{smart_progress}
\title{smart_progress}
\usage{
smart_progress(ii, nn, string = ".", iter = 5, iter2 = 200, ...)
}
\arguments{
\item{ii}{A positive integer to track a 
loop's progress}

\item{nn}{A positive integer for the 
total number of loop iterations}

\item{string}{A string to print}

\item{iter}{A positive integer for how many
multiple iterations to print "."}

\item{iter2}{A positive integer to end
a line of printed "." and track the loop's 
progress}

\item{...}{arguments passed to cat}
}
\description{
smart_progress
}
