% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{smart_RT}
\alias{smart_RT}
\title{smart_RT}
\usage{
smart_RT(...)
}
\arguments{
\item{...}{arguments passed to read.table}
}
\value{
Return is identical to \code{read.table()}
}
\description{
Calls \code{read.table()} but sets
argument \code{stringsAsFactors = FALSE} to prevent
treating character columns as factors.
}
