% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smarter.R
\name{smart_SN}
\alias{smart_SN}
\title{smart_SN}
\usage{
smart_SN(x, digits = 2)
}
\arguments{
\item{x}{A numeric vector to convert 
to scientific notation}

\item{digits}{A positive integer for 
number of digits to include in notation}
}
\description{
Convert numeric values into 
scientific notation
}
