% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{smart_table}
\alias{smart_table}
\title{smart_table}
\usage{
smart_table(...)
}
\arguments{
\item{...}{arguments passed to table}
}
\value{
Return a table
}
\description{
Should elements passed into \code{table()}
contain \code{NA} or \code{NaN}, we want to see them 
by default.
}
\examples{

aa = c(1,1,2,2,2,3,NA)
table(aa)
smart_table(aa)
}
