% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{smart_WT}
\alias{smart_WT}
\title{smart_WF}
\usage{
smart_WT(...)
}
\arguments{
\item{...}{arguments passed to write.table}
}
\value{
Return is identical to \code{write.table()}
}
\description{
Calls \code{write.table()} setting
parameters \code{row.names} and \code{quote} 
to \code{FALSE}.
}
