\name{date_filter_types_available_per_query_type}
\alias{query types in SMART}
\docType{data}
\title{Features of Query types in SMART}
\description{
  A data frame with 26 observations and 10 variables
}
\usage{date_filter_types_available_per_query_type}
\format{
  A data frame with 26 observations and 10 variables
  }

\keyword{datasets}
