% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_conf_model.R
\name{flat_conf_model}
\alias{flat_conf_model}
\title{Flat table of a Configurable Model in a specified language}
\usage{
flat_conf_model(path_conf_model, language_interest = "en")
}
\arguments{
\item{path_conf_model}{the path to the xml file ending in "/xml_file_name.xml"}

\item{language_interest}{a string code with the language of interest. If the
configurable model has more than one language, use this function once per
language. English is the default language ("en"). If you want the English version
of the configurable model, jsut provide the path_conf_model. If the configurable
model also has Laotian language then provide the string "lo". For other languages
see:
https://meta.wikimedia.org/wiki/Template:List_of_language_names_ordered_by_code}
}
\value{
A tibble. The tibble has the following
columns:
"cat_key", "cat_label", "att_type", "att_key", "att_label",
"root_key", "root_label", "att_option_key", and "att_option_label". The "cat",
"att", "root" strings in the column headers refer to Category, Attribute,
and Tree data, respectively. The "key", "label", and "option" strings refer to
data regarding the key, the label, and the options available, respectively.
The output only contains active Categories, Attributes, and Options. Keys do
not change across languages.
}
\description{
Function to create a flat table from a SMART Configurable Model downloaded
from a Conservation Area in SMART Desktop exported as xml. The output has all
categories, attributes (list, multilist, trees, and others), and options.
Keys and labels for each item are provided. All inactive categories, attributes,
and options are NOT included in the result.
}
\details{
Currently, it is not possible to obtain the configurable model directly from
SMART Connect, so the Configurable Model must be exported from the Conservtion
Areas in SMART Desktop as an xml file to a known destination in your computer.
Then provide the path to the xml file in your local computer in the 'path_conf_model'
attribute of this function.
To export a Configurable Model, open SMART Desktop and the corresponding
Conservation Area. Go to the Conservation Area tab and click on "Configurable
Model". Select the Configurable Model you would like to present as a flat
table from the list and then click on "Export". Choose the location to save the
xml file in your computer. Then run this function providing the corresponding
path.

Tree attributes of the configurable model are assumed to have roots and options
(two levels top).
}
\examples{
flat_conf_model("your_path/xml_file_name.xml", language_interest)

}
