% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_info.R
\name{query_info}
\alias{query_info}
\title{Properties of the queries available in SMART Connect}
\usage{
query_info(server_url, user, password)
}
\arguments{
\item{server_url}{A string with the URL of the SMART Connect server
(e.g., "https://wcshealth.smartconservationtools.org/server" )}

\item{user}{A string with corresponding SMART Connect username
(e.g., "connect_username")}

\item{password}{A string with the corresponding SMART Connect user's password
(e.g., "my_connect_password")}
}
\value{
A list with nested tibbles. The length of the list corresponds to the number
SMART conservation areas the user has access to. The number nested tibbles
per each list object corresponds to the number of queries available per
conservation area. Each nested tibble has the information of a query
shown in two columns: "feature" and "value".
The features are: folder, query_name, typeKey, uuid,  spatial_query,
date_filter_type:assetdeploymentdate, date_filter_type:missionenddate,
date_filter_type:missionstartdate, date_filter_type:missiontrackdate,
date_filter_type:patrolend, date_filter_type:patrolstart,
date_filter_type:RecordDate, date_filter_type:waypointdate,
date_filter_type:waypointlastmodified. See the details section.
The objects are named so it is possible to find the information of a specific
query with list$conservation_area_name$query_name
}
\description{
A function to learn about the queries available in SMART Connect and their
properties. This information is relevant to provide the correct parameters
for the function 'data_from_connect' and load the corresponding query output
in the R session.
}
\details{
Not all SMART queries have the same properties. The 'query_info function'
provides the properties of the queries in each conservation area
following the permissions of SMART Connect users.
'folder' column provides the
query folder of the conservation area where the query is stored.
query_name' provides the query name.
'typeKey' provides the type of query.
'uuid' is an internal SMART identifier for the query
'spatial_query' is a column with T/F values where TRUE means
the query has spatial information and otherwise if FALSE.
The 'date_filter_type:' features are also T/F showing the
available temporal filters options of the query. For example, a
'date_filter_type:waypointdate' equals TRUE means that 'waypointdate' can
be chosen as the reference for the start and end dates of the data to
be included in the output provided by the function data_from_connect'.
}
\examples{
user<-"your_smart_connect_user_name"
password<-"your_smart_connect_password"
server_url<-"https://wcshealth.smartconservationtools.org/server"
query_data<-query_info(user=user, password=password, server_url=server_url)


}
\seealso{
\link{data_from_connect}
}
