% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_info.R
\name{query_info}
\alias{query_info}
\title{#' Explore the queries available in SMART Connect}
\usage{
query_info(server_url, user, password)
}
\arguments{
\item{server_url}{A string with the URL of the SMART Connect server
(e.g., "https://wcshealth.smartconservationtools.org/server" )}

\item{user}{A string with corresponding SMART Connect username
(e.g., "connect_username")}

\item{password}{A string with the corresponding SMART Connect user's password
(e.g., "my_connect_password")}
}
\value{
A tibble with the columns ....
}
\description{
#' Explore the queries available in SMART Connect
}
\details{
Not all SMART queries have spatial information included and different query
types have different properties including day filter options. To learn about
the Conservation Areas available in SMART Connect, the queries available per
Conservation Area, and their properties such as query type, spatial
information availability (the type_output parameter of this function), and
date filter options (the date_filter parameter of this function) use the
"queries_available_per_conservation_area" function first.
}
\examples{
x<-1

}
\seealso{
\code{\link[=queries_available_per_conservation_area]{queries_available_per_conservation_area()}} to learn about the Conservation
Areas available in SMART Connect, the queries in each Conservation Area, the
type of each query, spatial data in the queries, and the options available to
filter by date for each query.
}
