% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartr.R
\name{smart_merge}
\alias{smart_merge}
\title{smart_merge}
\usage{
smart_merge(x, y, mess = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}

\item{y}{A data.frame}

\item{mess}{Default to \code{FALSE}.
Otherwise a message is printed.}

\item{...}{arguments passed to merge}
}
\description{
smart_merge
}
