% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartr.R
\name{smart_reqNames}
\alias{smart_reqNames}
\title{smart_reqNames}
\usage{
smart_reqNames(DATA, REQ)
}
\arguments{
\item{DATA}{A matrix or data.frame}

\item{REQ}{A string vector of colnames
required to be contained in DATA}
}
\description{
smart_reqNames
}
