% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{smart_rmcols}
\alias{smart_rmcols}
\title{smart_rmcols}
\usage{
smart_rmcols(OBJ, rm_names)
}
\arguments{
\item{OBJ}{A matrix or data.frame}

\item{rm_names}{A string vector of colnames 
to remove}
}
\description{
smart_rmcols
}
