% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartr.R
\name{smart_table}
\alias{smart_table}
\title{smart_table}
\usage{
smart_table(...)
}
\arguments{
\item{...}{arguments passed to table}
}
\description{
smart_table
}
\examples{

aa = c(1,1,2,2,2,3,NA)
table(aa)
smart_table(aa)

}
